+++ tools/testdata/gen_proto_grpc_runtime/testdata/protogrpce2eprobe/main.go
@@ -16,7 +16,7 @@ func main() {
 	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
 	defer cancel()
 
-	conn, err := grpc.DialContext(ctx, "127.0.0.1:10621",
+	conn, err := grpc.DialContext(ctx, "127.0.0.1:12310",
 		grpc.WithTransportCredentials(insecure.NewCredentials()),
 		grpc.WithBlock(),
 	)